refactor(search): :truck: move CountElement to ResultCount
