+++ oarepo_ui/theme/assets/semantic-ui/js/oarepo_ui/search/index.js
@@ -3,7 +3,7 @@ export { ActiveFiltersElement } from "./ActiveFiltersElement";
 export { BucketAggregationElement } from "./BucketAggregationElement";
 export { BucketAggregationValuesElement } from "./BucketAggregationValuesElement";
 
-export { ResultCountWithState } from "./CountElement";
+export { ResultCountWithState, ResultCount } from "./ResultCount";
 export { ErrorElement } from "./ErrorElement";
 
 export { SearchAppFacets } from "./SearchAppFacets";
